% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_time_series.R
\name{add_time_series}
\alias{add_time_series}
\title{Add time series to a multiple time series plot.}
\usage{
add_time_series(
  h = NULL,
  f = NULL,
  r = NULL,
  t = NULL,
  mt = NULL,
  mc = NULL,
  colour = "red"
)
}
\arguments{
\item{h}{A data frame containing the spatial coordinates representing the
marks in an additional humerus sample.}

\item{f}{A data frame containing the spatial coordinates representing the
marks in an additional femur sample.}

\item{r}{A data frame containing the spatial coordinates representing the
marks in an additional radius sample.}

\item{t}{A data frame containing the spatial coordinates representing the
marks in an additional tibia sample.}

\item{mt}{A data frame containing the spatial coordinates representing the
marks in an additional metatarsus sample.}

\item{mc}{A data frame containing the spatial coordinates representing the
marks in an additional metacarpus sample.}

\item{colour}{A lower case string specifying the colour of points and lines
in the plot to represent the additional set of samples.}
}
\value{
A data frame including the sequence and frequencies of all the
additional samples. Mark frequencies are added to the plot in the popup
window previously created with the \code{create_time_series} function.
}
\description{
The present function is used to add a time series plot to an
already existing time series plot.
}
\examples{
data("humerus_right_circular1")
data("femur_right_circular1")
data("radius_right_circular1")
data("tibia_right_circular1")
h1 <- load_marks(femur_right_circular1, mark_type = "circular", plot = FALSE)
f1 <- load_marks(femur_right_circular1, mark_type = "circular", plot = FALSE)
r1 <- load_marks(femur_right_circular1, mark_type = "circular", plot = FALSE)
t1 <- load_marks(femur_right_circular1, mark_type = "circular", plot = FALSE)
time_series_1 <- create_time_series(h = h1, f = f1, r = r1, t = t1)

data("humerus_right_circular2")
data("femur_right_circular2")
data("radius_right_circular2")
data("tibia_right_circular2")
h2 <- load_marks(femur_right_circular2, mark_type = "circular", plot = FALSE)
f2 <- load_marks(femur_right_circular2, mark_type = "circular", plot = FALSE)
r2 <- load_marks(femur_right_circular2, mark_type = "circular", plot = FALSE)
t2 <- load_marks(femur_right_circular2, mark_type = "circular", plot = FALSE)
time_series_2 <- add_time_series(h = h2, f = f2, r = r2, t = t2, colour = "red")

data("humerus_right_circular3")
data("femur_right_circular3")
data("radius_right_circular3")
data("tibia_right_circular3")
h3 <- load_marks(femur_right_circular3, mark_type = "circular", plot = FALSE)
f3 <- load_marks(femur_right_circular3, mark_type = "circular", plot = FALSE)
r3 <- load_marks(femur_right_circular3, mark_type = "circular", plot = FALSE)
t3 <- load_marks(femur_right_circular3, mark_type = "circular", plot = FALSE)
time_series_3 <- add_time_series(h = h3, f = f3, r = r3, t = t3, colour = "blue")
}
\seealso{
\code{\link{create_time_series}}, \code{\link{load_marks}}.
}
\author{
Lloyd A. Courtenay
}
