% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_orientations.R
\name{calculate_orientations}
\alias{calculate_orientations}
\title{Calculation of orientation of linear marks.}
\usage{
calculate_orientations(linear_data, bone, create_external_plot = TRUE)
}
\arguments{
\item{linear_data}{A string containing the path of the Ikhnos product file
containing linear data (e.g., scores or cut marks). If no file name is
specified, then a popup window will appear allowing the user to locate the
file in their system.}

\item{bone}{An IkhnosToolBox data object (e.g., data(left_femur)) containing
the 3d model of the bone being studied.}

\item{create_external_plot}{A boolean TRUE or FALSE (default = TRUE) option
to create two popup windows; one for the visualisation of the location of the
linear marks on the specified bone; and a second one with the rose diagram
based on the orientations of the marks.}
}
\value{
A \code{circular} object containing the angles calculated between
the bone long axis and each linear mark. If create_external_plot = TRUE, then
two popup windows are created for the rose diagram and the visualisation
of the linear marks on the specified bone.
}
\description{
Function to conduct orientation analyses of scores and/or cut
marks.
}
\section{Bibliography}{

Edelsbrunner, H., Kirkpatrick, D.G. and Seidel, R. (1983). On the shape of a
set of points in the plane. IEEE Transactions on Information Theory, 29(4), pp.551-559.

Jammalamadaka, S. Rao and SenGupta, A. (2001). Topics in Circular Statistics,
World Scientific Press, Singapore.
}

\examples{
data("right_femur")
data("femur_right_linear1")
data("femur_right_linear2")
example_circ_1 <- load_marks(femur_right_linear1, mark_type = "linear")
example_circ_2 <- load_marks(femur_right_linear2, mark_type = "linear")
example_calculate_orientations_1 <- calculate_orientations(example_circ_1, right_femur)
example_calculate_orientations_2 <- calculate_orientations(example_circ_2, right_femur)
}
\seealso{
\code{\link{load_marks}}, \code{\link{ashape}}, \code{\link{circular}},
\code{\link{arrows.circular}}
}
\author{
Lloyd A. Courtenay
}
