% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_two_sample_orientations.R
\name{compare_two_sample_orientations}
\alias{compare_two_sample_orientations}
\title{Pair comparison for homogeneity on two orientation patterns.}
\usage{
compare_two_sample_orientations(sample_1, sample_2)
}
\arguments{
\item{sample_1}{A circular object containing the angles calculated between
the bone long axis and each linear mark in the first dataset.}

\item{sample_2}{A circular object containing the angles calculated between
the bone long axis and each linear mark in the second dataset.}
}
\value{
A list containing the results obtained for the Mardia-Watson-Wheeler
test and the associated p value.
}
\description{
Function to compare the orientation patterns observed in two
different samples of linear marks (scores and/or cut marks)using the
randomised Mardia-Watson-Wheeler test.
}
\section{Bibliography}{

Jammalamadaka, S. Rao and SenGupta, A. (2001). Topics in Circular Statistics,
World Scientific Press, Singapore.
}

\examples{
data("femur_right_linear1")
data("femur_right_linear2")
data("right_femur")
example_circ_1 <- load_marks(femur_right_linear1, mark_type = "linear")
example_circ_2 <- load_marks(femur_right_linear2, mark_type = "linear")
example_calculate_orientations_1 <- calculate_orientations(
  example_circ_1, right_femur
)
example_calculate_orientations_2 <- calculate_orientations(
  example_circ_2, right_femur
)
example_comp_orientations <- compare_two_sample_orientations(
  example_calculate_orientations_1, example_calculate_orientations_2
)
}
\seealso{
\code{\link{calculate_orientations}}, \code{\link{circular}}.
}
\author{
Lloyd A.Courtenay
}
