% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_time_series.R
\name{create_time_series}
\alias{create_time_series}
\title{Creation of multiple time series plot.}
\usage{
create_time_series(
  h = NULL,
  f = NULL,
  r = NULL,
  t = NULL,
  mt = NULL,
  mc = NULL,
  colour = "black",
  create_external_plot = TRUE
)
}
\arguments{
\item{h}{A data frame containing the spatial coordinates representing the
marks in the first humerus sample.}

\item{f}{A data frame containing the spatial coordinates representing the
marks in the first femur sample.}

\item{r}{A data frame containing the spatial coordinates representing the
marks in the first radius sample.}

\item{t}{A data frame containing the spatial coordinates representing the
marks in the first tibia sample.}

\item{mt}{A data frame containing the spatial coordinates representing the
marks in the first metatarsus sample.}

\item{mc}{A data frame containing the spatial coordinates representing the
marks in the first metacarpus sample.}

\item{colour}{A lower case string specifying the colour of points and lines
in the plot to represent the first set of samples.}

\item{create_external_plot}{A boolean TRUE or FALSE (default = TRUE) option
to create a popup window with the plot for mark frequency for all the samples.}
}
\value{
A data frame including the sequence and frequencies of all the samples.
If create_external_plot = TRUE, then the function returns a popup
window with a single plot showing the mark frequencies for all the elements.
}
\description{
The present function is used to generate a plot including the
mark frequencies on several bone elements and provide the basis for the
addition of further samples.
}
\examples{
#DA PROBLEMAS CON mt - HE VISTO UNOS FALLOS EN MISC FUNCTIONS: PONIA MC CUANDO DEBER?A PONER MT, LO HE CAMBIADO Y SIGUE SIN FUNCIONAR - EL RESTO DE HUESOS FUNCIONA BIEN
data(humerus_right_circular1)
data(femur_right_circular1)
data(radius_right_circular1)
data(tibia_right_circular1)  #COMPROBAR ESTO
h1 <- load_marks(femur_right_circular1, mark_type = "circular", plot = FALSE) #ESTO DA ERROR POR NO CARGAR EL RDA
f1 <- load_marks(femur_right_circular1, mark_type = "circular", plot = FALSE) #ESTO DA ERROR POR NO CARGAR EL RDA
r1 <- load_marks(femur_right_circular1, mark_type = "circular", plot = FALSE) #ESTO DA ERROR POR NO CARGAR EL RDA
t1 <- load_marks(femur_right_circular1, mark_type = "circular", plot = FALSE) #ESTO DA ERROR POR NO CARGAR EL RDA
time_series_1 <- create_time_series(h = h1, f = f1, r = r1, t = t1)
}
\seealso{
\code{\link{add_time_series}}, \code{\link{load_marks}}.
}
