% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_circular_analysis.R
\name{descriptive_circular_analysis}
\alias{descriptive_circular_analysis}
\title{Descriptive circular analysis.}
\usage{
descriptive_circular_analysis(circular_object)
}
\arguments{
\item{circular_object}{A circular object containing the angles calculated
between the bone long axis and each linear mark.}
}
\value{
A list with the results obtained for the standarised measures of
Skewness and Kurtosis, the sample circular variance and dispersion,
and the central orientation radians and degrees.
}
\description{
The present function contains several parameters to characterise
the location and variability of linear marks along the bone, including
normality distribution tests such as skewness (measure of asymmetry) and
kurtosis (measure of tailedness and peakedness), circular variance and
dispersion, and the orientation of the radians and degrees.
}
\section{Bibliography}{

Jammalamadaka, S. Rao and SenGupta, A. (2001). Topics in Circular Statistics,
World Scientific Press, Singapore.
}

\examples{
data("femur_right_linear1")
data("femur_right_linear2")
data("right_femur")
example_circ_1 <- load_marks(femur_right_linear1, mark_type = "linear")
example_circ_2 <- load_marks(femur_right_linear2, mark_type = "linear")
example_calculate_orientations_1 <- calculate_orientations(
  example_circ_1, right_femur,
  create_external_plot = FALSE
)
example_calculate_orientations_2 <- calculate_orientations(
  example_circ_2, right_femur,
  create_external_plot = FALSE
)
example_circ_analysis1 <- descriptive_circular_analysis(example_calculate_orientations_1)
example_circ_analysis2 <- descriptive_circular_analysis(example_calculate_orientations_2)
}
\seealso{
\code{\link{calculate_orientations}}, \code{\link{circular}},
\code{\link{trigonometric.moment}}.
}
\author{
Lloyd A.Courtenay
}
