% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_spatial_data.R
\name{extract_spatial_data}
\alias{extract_spatial_data}
\title{Prepare data for spatial analysis}
\usage{
extract_spatial_data(
  input_data,
  mark_type = "circular",
  print_summary = TRUE,
  plot_results = FALSE
)
}
\arguments{
\item{input_data}{a data.frame object containing spatial coordinates}

\item{mark_type}{a string specifying whether marks are "circular" or "linear"}

\item{print_summary}{a boolean option to print a summary of data}

\item{plot_results}{a boolean option to plot the spatial information}
}
\value{
A spatial object for statistical analyses
}
\description{
The present function is used to prepare data for further spatial analysis
}
\examples{
example_data <- load_marks(mark_type = "circular")
example_sp_object <- extract_spatial_data(example_data, "circular")
}
