% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IkhnosToolBox.support.code.r
\docType{data}
\name{femur_right_circular3}
\alias{femur_right_circular3}
\title{x, y and z coordinates for pits on right small-sized ungulate femora}
\description{
x, y and z coordinates for pits on right small-sized ungulate femora
}
\references{
in prep
}
\author{
Julia Aramendi
}
\keyword{datasets}
