% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IkhnosToolBox.support.code.r
\docType{data}
\name{humerus_right_circular1}
\alias{humerus_right_circular1}
\title{x, y and z coordinates for pits on right large-sized ungulate humeri}
\description{
x, y and z coordinates for pits on right large-sized ungulate humeri
}
\references{
in prep
}
\author{
Julia Aramendi
}
\keyword{datasets}
