% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IkhnosToolBox_support.r
\docType{data}
\name{humerus_right_circular2}
\alias{humerus_right_circular2}
\title{x, y and z coordinates for pits on right medium-sized ungulate humeri}
\description{
x, y and z coordinates for pits on right medium-sized ungulate humeri
}
\references{
in prep
}
\author{
Julia Aramendi
}
\keyword{datasets}
