% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IkhnosToolBox_support.r
\docType{package}
\name{IkhnosToolBox}
\alias{IkhnosToolBox}
\title{R Tool Box for the Ikhnos Taphonomic Software}
\description{
Group of R functions for the statistical analysis of data derived from the
Ikhnos taphonomic software. Functions in this package allow one to visualise the
location and orientation of bone surface modifications such as tooth and percussion pits,
cut marks and tooth scores on long bones and their analysis for the identification and
comparison of orientation and distribution patterns along the bone surface. Functions
are available to perform analyses within samples as well as across different samples.
}
\author{
Lloyd A. Courtenay and Julia Aramendi
}
