% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IkhnosToolBox_support.r
\docType{data}
\name{left_humerus}
\alias{left_humerus}
\title{3D model of cervid left humerus}
\description{
3D model of cervid left humerus
}
\details{
sample property rights-Jose Yravedra
}
\references{
in prep
}
\author{
Julia Aramendi
}
\keyword{3D}
\keyword{model}
