% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IkhnosToolBox_support.r
\docType{data}
\name{left_metatarsus}
\alias{left_metatarsus}
\title{3D model of cervid left metatarsus}
\description{
3D model of cervid left metatarsus
}
\details{
sample property rights-Jose Yravedra
}
\references{
in prep
}
\author{
Julia Aramendi
}
\keyword{3D}
\keyword{model}
