% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_marks.R
\name{load_marks}
\alias{load_marks}
\title{Load marks from IKHNOS files.}
\usage{
load_marks(
  data,
  file_name = NULL,
  mark_type = "circular",
  delim = ",",
  plot = FALSE,
  colour_value = "black",
  mark_size = 6
)
}
\arguments{
\item{file_name}{A string containing the path of the Ikhnos product file.
If no file name is specified, then a popup window will appear allowing the
user to locate the file in their system}

\item{mark_type}{A lower case string specifying whether the marks under study
are "circular" or "linear".}

\item{delim}{A string specifying how marks are delimited within the Ikhnos
product file.}

\item{plot}{A boolean TRUE or FALSE option (default = FALSE) defining whether
to produce a 3D plot of the marks.}

\item{colour_value}{A lower case string specifying the colour of the marks in
the case where plot = TRUE.}

\item{mark_size}{The size of the points or lines produced in the 3D plot.}
}
\value{
A \code{data.frame} object containing the data extracted from Ikhnos.
If plot = TRUE, then a 3D visualisation of marks is also created.
}
\description{
The present function allows the user to locate a file and load
data derived from Ikhnos for both visualisation and preparation for
statistical applications. If the \code{load_bone} function has already been
used, then marks are visualised directly on the surface of the bone that is
currently in view. If no \code{load_bone} function has been called, then the
plot will contain only the 3D distribution of marks.
This can then be combined with functions such as \code{load_marks}
to visualise marks on the 3D model, and \code{save_3d_image} to
save the window to a *.png file.
}
\examples{
data("right_femur")
data("femur_right_circular1")
data("femur_right_linear1")
load_bone(right_femur)
load_marks(femur_right_circular1, mark_type = "circular", plot = TRUE, colour_value = "black")
load_marks(femur_right_linear1, mark_type = "linear", plot = TRUE, colour_value = "red")
save_3d_image("my_first_plot")
}
\seealso{
\code{\link{load_bone}}, \code{\link{save_3d_image}},
\code{\link{points3d}}
}
