% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IkhnosToolBox_support.r
\docType{data}
\name{metatarsus_right_circular1}
\alias{metatarsus_right_circular1}
\title{x, y and z coordinates for pits on right large-sized ungulate metatarsal bones}
\description{
x, y and z coordinates for pits on right large-sized ungulate metatarsal bones
}
\references{
in prep
}
\author{
Julia Aramendi
}
\keyword{datasets}
