% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IkhnosToolBox.support.code.r
\docType{data}
\name{right_humerus}
\alias{right_humerus}
\title{3D model of cervid right humerus}
\description{
3D model of cervid right humerus
}
\references{
in prep
}
\author{
Julia Aramendi
}
\keyword{3D}
\keyword{model}
