% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IkhnosToolBox_support.r
\docType{data}
\name{right_tibia}
\alias{right_tibia}
\title{3D model of cervid right tibia}
\description{
3D model of cervid right tibia
}
\references{
in prep
}
\author{
Julia Aramendi
}
\keyword{3D}
\keyword{model}
