% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_3d_image.R
\name{save_3d_image}
\alias{save_3d_image}
\title{Saves 3D visualization.}
\usage{
save_3d_image(file_name)
}
\arguments{
\item{file_name}{a string specifying the file name where the .png image will be saved.}
}
\value{
Saves the visualized 3D window as a .png file.
}
\description{
Function to save visualization of data.
}
\examples{
data(left_femur)
load_bone(left_femur)
load_marks(mark_type = "circular", plot = TRUE, colour_value = "black")
load_marks(mark_type = "linear", plot = TRUE, colour_value = "red")
save_3d_image("my_first_plot")
}
\seealso{
\code{\link{load_bone}}, \code{\link{save_3d_image}}.
}
