% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_distance_matrices.R
\name{sp_distance_matrices}
\alias{sp_distance_matrices}
\title{Comparison of nearest neighbour distance matrices.}
\usage{
sp_distance_matrices(
  spatial_object_1,
  spatial_object_2,
  name_1,
  name_2,
  create_external_plot = TRUE
)
}
\arguments{
\item{spatial_object_1}{A pp3 object containing a 3-dimensional point pattern.}

\item{spatial_object_2}{A pp3 object containing a second 3-dimensional point
pattern.}

\item{name_1}{A string to define the title of the plot that corresponds to
\code{spatial_object_1}.}

\item{name_2}{A string to define the title of the plot that corresponds to
\code{spatial_object_2}.}

\item{create_external_plot}{A boolean TRUE or FALSE (default = TRUE) option
to create two popup windows with the plots of the distances between the marks
within and between each sample.}
}
\value{
If create_external_plot = TRUE, then the function returns two popup
windows with:
1. two plots for the calculated nearest neighbour distances between marks
within each sample
2. one plot for the calculated nearest neighbour distances between marks
between the two samples
}
\description{
The present function is used to calculate the nearest neighbour
distance within and between two samples.
}
\section{Bibliography}{

A. Baddeley, E. Rubak and R.Turner. Spatial Point Patterns: Methodology and
Applications with R. Chapman and Hall/CRC Press, 2015.
}

\examples{
data("femur_right_circular1")
data("femur_right_linear1")
example_data1 <- load_marks(femur_right_circular1, mark_type = "circular")
example_data2 <- load_marks(femur_right_linear1, mark_type = "linear")
example_sp_object1 <- extract_spatial_data(example_data1, "circular")
example_sp_object2 <- extract_spatial_data(example_data2, "linear")
example_distance_matrices <- sp_distance_matrices(
  example_sp_object1, example_sp_object2,
  name_1 = "circular", name_2 = "linear"
)
}
\seealso{
\code{\link{sp_distance_matrix}}, \code{\link{crossdist}},
\code{\link{nndist}}
}
