% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IkhnosToolBox.support.code.r
\docType{data}
\name{tibia_right_circular1}
\alias{tibia_right_circular1}
\title{x, y and z coordinates for pits on right large-sized ungulate tibiae}
\description{
x, y and z coordinates for pits on right large-sized ungulate tibiae
}
\references{
in prep
}
\author{
Julia Aramendi
}
\keyword{datasets}
