% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_clustering.R
\name{time_series_clustering}
\alias{time_series_clustering}
\title{Cluster analysis.}
\usage{
time_series_clustering(series_dataframe, create_external_plot = TRUE)
}
\arguments{
\item{series_dataframe}{A data frame including the sequence and frequencies
per sample.}

\item{create_external_plot}{A boolean TRUE or FALSE (default = TRUE) option
to create a popup window with the plot for the cluster analysis.}
}
\value{
If create_external_plot = TRUE, then the function returns a popup
window with a plot on the cluster analysis.
}
\description{
The present function is used to conduct a cluster analysis on
several samples.
}
\examples{
data(humerus_right_circular1)
data(femur_right_circular1)
data(radius_right_circular1)
data(tibia_right_circular1)  #COMPROBAR ESTO
h1 <- load_marks(femur_right_circular1, mark_type = "circular", plot = FALSE) #ESTO DA ERROR POR NO CARGAR EL RDA
f1 <- load_marks(femur_right_circular1, mark_type = "circular", plot = FALSE)
r1 <- load_marks(femur_right_circular1, mark_type = "circular", plot = FALSE)
t1 <- load_marks(femur_right_circular1, mark_type = "circular", plot = FALSE)
time_series_1 <- create_time_series(h = h1, f = f1, r = r1, t = t1)

data(humerus_right_circular2)
data(femur_right_circular2)
data(radius_right_circular2)
data(tibia_right_circular2)
h1 <- load_marks(femur_right_circular2, mark_type = "circular", plot = FALSE)
f1 <- load_marks(femur_right_circular2, mark_type = "circular", plot = FALSE)
r1 <- load_marks(femur_right_circular2, mark_type = "circular", plot = FALSE)
t1 <- load_marks(femur_right_circular2, mark_type = "circular", plot = FALSE)
time_series_2 <- create_time_series(h = h2, f = f2, r = r2, t = t2, colour = "red")

data(humerus_right_circular3)
data(femur_right_circular3)
data(radius_right_circular3)
data(tibia_right_circular3)
h1 <- load_marks(femur_right_circular3, mark_type = "circular", plot = FALSE)
f1 <- load_marks(femur_right_circular3, mark_type = "circular", plot = FALSE)
r1 <- load_marks(femur_right_circular3, mark_type = "circular", plot = FALSE)
t1 <- load_marks(femur_right_circular3, mark_type = "circular", plot = FALSE)
time_series_3 <- create_time_series(h = h3, f = f3, r = r3, t = t3, colour = "blue")

series_database <- rbind(time_series_1, time_series_2, time_series_3)
series_database$Sample <- as.factor(c(rep("Series_1", nrow(time_series_1)),rep("Series_2", nrow(time_series_2)),rep("Series_3", nrow(time_series_3))))
time_series_clustering(series_database)
Warning message:`guides(<scale> = FALSE)` is deprecated. Please use `guides(<scale> = "none")` instead.
}
\seealso{
\code{\link{create_time_series}}, \code{\link{add_time_series}}.
}
