% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_sample_histogram_distributions.R
\name{two_sample_histogram_distributions}
\alias{two_sample_histogram_distributions}
\title{Create pair histograms.}
\usage{
two_sample_histogram_distributions(
  group_1,
  group_2,
  sample_1_name,
  sample_2_name,
  dimension,
  bone_type,
  create_external_plot = TRUE
)
}
\arguments{
\item{group_1}{A pp3 object containing a 3-dimensional point pattern.}

\item{group_2}{A pp3 object containing a second 3-dimensional point pattern.}

\item{sample_1_name}{A string to define the label of \code{group_1}.}

\item{sample_2_name}{A string to define the label of \code{group_2}.}

\item{dimension}{A character to define the axis;\code{x} refers to proximal-
distal axis,\code{y} to the lateral-medial axis, and \code{z} to the caudal-
cranial axis.}

\item{bone_type}{A string to define the element; possible bones are
\code{femur}, \code{humerus}, \code{metacarpus}, \code{radius}, \code{tibia},
\code{metatarsus}.}

\item{create_external_plot}{A boolean TRUE or FALSE option (default = TRUE)
to create a popup window with the plot for the distribution of both samples}
}
\value{
A \code{list} containing the number of breaks and counts of both
samples.
}
\description{
The present function is used to compare the frequency of marks
along a specific bone axis between two samples.
}
\section{Bibliography}{

Bowman, A.W. and Azzalini, A. (1997). Applied Smoothing Techniques for Data
Analysis: the Kernel Approach with S-Plus Illustrations. Oxford University Press, Oxford.
}

\examples{
data("femur_right_circular1")
data("femur_right_linear1")
example_data1 <- load_marks(femur_right_circular1, mark_type = "circular")
example_data2 <- load_marks(femur_right_linear1, mark_type = "linear")
example_sp_object1 <- extract_spatial_data(example_data1, "circular")
example_sp_object2 <- extract_spatial_data(example_data2, "circular")
example_time_series <- two_sample_histogram_distributions(
  example_sp_object1, example_sp_object2,"circular", "linear","x", "femur",
  create_external_plot = FALSE
)
}
\seealso{
\code{\link{wavelet_analysis}}, \code{\link{sm.density.compare}}.
}
\author{
Lloyd A.courtenay
}
