% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_spatial_data.R
\name{extract_spatial_data}
\alias{extract_spatial_data}
\title{Prepare data for spatial analysis.}
\usage{
extract_spatial_data(
  input_data,
  mark_type = "circular",
  print_summary = TRUE,
  plot_results = FALSE,
  create_external_plot = TRUE
)
}
\arguments{
\item{input_data}{A data frame containing spatial coordinates.}

\item{mark_type}{A string specifying whether marks are "circular" or "linear".}

\item{print_summary}{A boolean TRUE or FALSE option (default = TRUE) to print
a summary of the data included in the spatial object.}

\item{plot_results}{A boolean TRUE or FALSE option (default = FALSE) to plot
the spatial information.}

\item{create_external_plot}{A boolean TRUE or FALSE (default = TRUE) option
to create a popup window with the plot for mark frequency for all the samples.}
}
\value{
A spatial object for statistical analyses.
}
\description{
The present function is used to prepare data for further spatial
analysis.
}
\examples{
data("femur_right_circular1")
example_data <- load_marks(femur_right_circular1, mark_type = "circular")
example_sp_object <- extract_spatial_data(example_data, "circular")+
}
\seealso{
\code{\link{box3}}, \code{\link{pp3}}.
}
\author{
Lloyd A.courtenay
}
