% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IkhnosToolBox_support.r
\docType{data}
\name{right_metacarpus}
\alias{right_metacarpus}
\title{3D model of cervid right metacarpus}
\description{
3D model of cervid right metacarpus
}
\details{
sample property rights-Jose Yravedra
}
\references{
in prep
}
\author{
Julia Aramendi
}
\keyword{3D}
\keyword{model}
