% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_3d_image.R
\name{save_3d_image}
\alias{save_3d_image}
\title{Save 3D visualisation.}
\usage{
save_3d_image(file_name)
}
\arguments{
\item{file_name}{A string specifying the file name where the *.png
image will be saved.}
}
\value{
Saves the visualised 3D window as a *.png file.
}
\description{
Function to save visualisation of data.
}
\examples{
data("right_femur")
data("femur_right_circular1")
data("femur_right_linear1")
load_bone(right_femur)
load_marks(femur_right_circular1, mark_type = "circular", plot = TRUE, colour_value = "black")
load_marks(femur_right_linear1, mark_type = "linear", plot = TRUE, colour_value = "red")
save_3d_image("my_first_plot")
}
\seealso{
\code{\link{load_bone}}, \code{\link{save_3d_image}},
\code{\link{rgl.snapshot}}.
}
