% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_distance_matrix.R
\name{sp_distance_matrix}
\alias{sp_distance_matrix}
\title{Nearest Neighbour distance matrix.}
\usage{
sp_distance_matrix(spatial_object, name = "marks", create_external_plot = TRUE)
}
\arguments{
\item{spatial_object}{A pp3 object containing the 3-dimensional point pattern}

\item{name}{A string to define the title of the plot.}

\item{create_external_plot}{A boolean TRUE or FALSE option (default = TRUE)
to create a popup window with the plot of the distances between the marks.}
}
\value{
If create_external_plot = TRUE, then the function returns an popup
window with the plot for the calculated nearest neighbour distances between
marks within the sample.
}
\description{
The present function is used to calculate the distances between
the marks in the sample.
}
\section{Bibliography}{

A. Baddeley, E. Rubak and R.Turner. Spatial Point Patterns: Methodology and
Applications with R. Chapman and Hall/CRC Press, 2015.
}

\examples{
data("femur_right_circular1")
example_data <- load_marks(femur_right_circular1, mark_type = "circular")
example_sp_object <- extract_spatial_data(example_data, "circular")
example_distance_matrix <- sp_distance_matrix(example_sp_object, name = "marks")
}
\seealso{
\code{\link{sp_distance_matrices}}, \code{\link{nndist}}
}
