% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_analysis.R
\name{wavelet_analysis}
\alias{wavelet_analysis}
\title{Bivariate wavelet analyses in frequency domains.}
\usage{
wavelet_analysis(
  first_ts,
  second_ts,
  sample_1_name,
  sample_2_name,
  n_iterations = 10000,
  x_lab = "Inter-epiphysis distance",
  create_external_plot = FALSE
)
}
\arguments{
\item{first_ts}{A matrix (n rows x 2 columns) containing the time series of
the first sample. The first column contains the time steps (breaks) and the
second column contains the values (counts).}

\item{second_ts}{A matrix (n rows x 2 columns) containing the time series of
the second sample. The first column contains the time steps (breaks) and the
second column contains the values (counts).}

\item{sample_1_name}{A string to define the label that corresponds to
\code{first_ts}.}

\item{sample_2_name}{A string to define the label that corresponds to
\code{second_ts}.}

\item{n_iterations}{A positive, non-zero integer to define the number of
Monte Carlo randomisation (default = 10000).}

\item{x_lab}{A string to define the title of the x axis.It should coincide
with the dimension selected for the \code{two_sample_histogram_distributions}
function.}

\item{create_external_plot}{A boolean TRUE or FALSE (default = TRUE) option
to create a popup window with the plot for the wavelet coherence analysis.}
}
\value{
If create_external_plot = TRUE, then the function returns a popup
window with a biwavelet plot to visualise the dependence between
\code{first_ts} and \code{second_ts}.
}
\description{
The present function is used to conduct wavelet coherence
analyses to assess the dependence existing between two mark patterns.
}
\section{Bibliography}{

Cazelles, B., M. Chavez, D. Berteaux, F. Menard, J. O. Vik, S. Jenouvrier, and
N. C. Stenseth. 2008. Wavelet analysis of ecological time series. Oecologia 156:287-304.

Grinsted, A., J. C. Moore, and S. Jevrejeva. 2004. Application of the cross
wavelet transform and wavelet coherence to geophysical time series. Nonlinear
Processes in Geophysics 11:561-566.

Liu, Y., X. San Liang, and R. H. Weisberg. 2007. Rectification of the Bias in
the Wavelet Power Spectrum. Journal of Atmospheric and Oceanic Technology 24:2093-2102.

Torrence, C., and G. P. Compo. 1998. A Practical Guide to Wavelet Analysis.
Bulletin of the American Meteorological Society 79:61-78.

Torrence, C., and P. J. Webster. 1998. The annual cycle of persistence in the
El Nino/Southern Oscillation. Quarterly Journal of the Royal Meteorological Society 124:1985-2004.

Veleda, D., R. Montagne, and M. Araujo. 2012. Cross-Wavelet Bias Corrected by
Normalizing Scales. Journal of Atmospheric and Oceanic Technology 29:1401-1408.
}

\examples{
data("femur_right_circular1")
data("femur_right_linear1")
example_data1 <- load_marks(femur_right_circular1, mark_type = "circular")
example_data2 <- load_marks(femur_right_linear1, mark_type = "linear")
example_sp_object1 <- extract_spatial_data(example_data1, "circular")
example_sp_object2 <- extract_spatial_data(example_data2, "circular")
example_time_series <- two_sample_histogram_distributions(
  example_sp_object1, example_sp_object2,"circular", "linear","x", "femur"
)
example_wavelet_analysis <- wavelet_analysis(
  example_time_series$first_sample_ts, example_time_series$second_sample_ts,
  "circular", "linear"
)
}
\seealso{
\code{\link{wtc}}, \code{\link{plot.biwavelet}},
\code{\link{two_sample_histogram_distributions}}.
}
